PaRSEC: thread-shift non-trivial release from message to worker threads
A message may contain an arbitrarily expensive release, which may
block PaRSEC's communication task for longer than desired.
Hence, if we detect more than one key or a reducer on the selected edge,
we thread-shift the handling to a worker thread by creating a high-priority
task. For split-metadata types, we also move the unpacking of the message
into a separate task as it may require starting one or more
data transfers. TTG uses a thread-safe version of the PaRSEC MPI backend
so this is fine and relieves the communication thread from
starting RMA communication.

This change also includes a shift to passing copy objects instead of object
references in an attempt to reduce the number of copies created
of received objects.

Signed-off-by: Joseph Schuchart <schuchart@icl.utk.edu>