Replace buffer tracking with serialization inspection
We have been tracking the buffers of structures by having the buffer
register itself with some metadata. This is clumsy and error-prone
(no support for nesting, for example). Now we use a specialized
archive to inspect all members of a value and apply some functor
to each buffer (or its internal parsec data).

This is also the first step towards a structure derived from
parsec_data_copy_t to better manage the life-time of the host copy.
We now only allow shared_ptr for non-owning buffers so some adjustments
must be made by users.

Signed-off-by: Joseph Schuchart <joseph.schuchart@stonybrook.edu>