Pass the Execution Space as TT template parameter
Defaults to Host execution so existing code is not affected.
Properly set by make_tt.

We cannot query TT::derivedT for flags because at the time TT
is instantiated because derivedT is incomplete at that point.
For now pass the Space as template parameter. We need to find
a different way if we want to have multiple implementations
of a task.

Signed-off-by: Joseph Schuchart <joseph.schuchart@stonybrook.edu>