Implement key generator-based broadcast
Instead of generating the keys inside the task and sending the data
on separate output terminals, we can send the task with the value
and generate the keys on the fly. The application registers a callback
that is used to query successor keys. This allows us to send the value
once instead of multiple times, once per output terminal.

We currently have to populate a tuple of vectors. Once Clang supports
std::generator we can revisit this and return generators from the callback.

Signed-off-by: Joseph Schuchart <joseph.schuchart@stonybrook.edu>